index: Make exception rethrow possible to opt out
For webkit / embedded browsers it can be useful to ignore
promise unhandled rejections and let the default behavior run.

For node the default behavior is terrible because it does not
fail hard, so we want this rethrow exception behavior for nodejs
to make the exception actually fail the node test process.